requirements.txt
python-dotenv
requests
pandas
psycopg2-binary
flask
flask-sqlalchemy
sqlalchemy
streamlit
matplotlib 
seaborn 
scikit-learn
simplejson
pytest
flask_sqlalchemy